LayoutTests/media/controls/airplay-picker-expected.txt

This tests the wireless device picker icon.

This test only runs in DRT!


EVENT: canplaythrough
PASS: We are using the apple idiom

Test the AppleTV picker button

PASS: Name is 'AppleTV Device Picker'
PASS: Has hidden class
PASS: CSS display: none
PASS: Aria label is 'Choose Wireless Display'

Simulate an available AppleTV

PASS: Name is 'AppleTV Device Picker'
PASS: Does not have hidden class
PASS: CSS display: block

Testing finished.
